requirements.txt
anthropic==0.79.0
python-dotenv==1.0.0
